// block-performance-1.0.0/src/library/url/make/type.ts
export interface Input {
    protocol: string;
    username?: string;
    password?: string;
    hostname: string;
    port?: string;
    pathname?: string;
    search_params?: { [key: string]: string };
    hash?: string;
}

export interface OutputOk {
    url: string;
}

export interface OutputFail {}

export interface Callback<R> {
    ok: (output: OutputOk) => R;
    fail: (err: Error) => R;
}
// ...
export type Output = OutputOk;